Refactor runtime Service Account generation
