+++ internal/controller/pkg/revision/runtime.go
@@ -1,6 +1,7 @@
 package revision
 
 import (
+	"github.com/crossplane/crossplane-runtime/pkg/errors"
 	"github.com/crossplane/crossplane-runtime/pkg/meta"
 	v1 "github.com/crossplane/crossplane/apis/pkg/v1"
 	"github.com/crossplane/crossplane/apis/pkg/v1alpha1"
@@ -18,13 +19,14 @@ const (
 )
 
 type RuntimeManifestBuilder struct {
-	controllerConfig *v1alpha1.ControllerConfig
-	runtimeConfig    *v1beta1.DeploymentRuntimeConfig
-	namespace        string
-	revision         v1.PackageWithRuntimeRevision
+	revision                  v1.PackageWithRuntimeRevision
+	namespace                 string
+	serviceAccountPullSecrets []corev1.LocalObjectReference
+	controllerConfig          *v1alpha1.ControllerConfig
+	runtimeConfig             *v1beta1.DeploymentRuntimeConfig
 }
 
-func NewRuntimeManifestBuilder(ctx context.Context, client client.Client, namespace string, pwr v1.PackageWithRuntimeRevision) (*RuntimeManifestBuilder, error) {
+func NewRuntimeManifestBuilder(ctx context.Context, client client.Client, namespace string, serviceAccount string, pwr v1.PackageWithRuntimeRevision) (*RuntimeManifestBuilder, error) {
 	b := &RuntimeManifestBuilder{
 		namespace: namespace,
 		revision:  pwr,
@@ -46,90 +48,20 @@ func NewRuntimeManifestBuilder(ctx context.Context, client client.Client, namesp
 		b.runtimeConfig = rc
 	}
 
+	sa := &corev1.ServiceAccount{}
+	// Fetch XP ServiceAccount to get the ImagePullSecrets defined there.
+	// We will append them to the list of ImagePullSecrets for the runtime
+	// ServiceAccount.
+	if err := client.Get(ctx, types.NamespacedName{Namespace: namespace, Name: serviceAccount}, sa); err != nil {
+		return nil, errors.Wrap(err, errGetServiceAccount)
+	}
+	b.serviceAccountPullSecrets = sa.ImagePullSecrets
+
 	return b, nil
 }
 
-type DeploymentOverrides func(deployment *appsv1.Deployment)
-
-func DeploymentWithName(name string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Name = name
-	}
-}
-
-func DeploymentWithNamespace(namespace string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Namespace = namespace
-	}
-}
-
-func DeploymentWithOwnerReferences(owners []metav1.OwnerReference) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.OwnerReferences = owners
-	}
-}
-
-func DeploymentWithSelectors(selectors map[string]string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Selector.MatchLabels = selectors
-		// Ensure that the pod template labels always contains the deployment
-		// selector.
-		for k, v := range selectors {
-			d.Spec.Template.Labels[k] = v
-		}
-	}
-}
-
-func DeploymentWithServiceAccount(sa string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Template.Spec.ServiceAccountName = sa
-	}
-}
-
-func DeploymentWithImagePullSecrets(secrets []corev1.LocalObjectReference) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Template.Spec.ImagePullSecrets = secrets
-	}
-}
-
-func DeploymentRuntimeWithImage(image string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Template.Spec.Containers[0].Image = image
-	}
-}
-
-func DeploymentRuntimeWithImagePullPolicy(policy corev1.PullPolicy) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Template.Spec.Containers[0].ImagePullPolicy = policy
-	}
-}
-
-func DeploymentRuntimeWithTLSServerSecret(secret string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		mountTLSSecret(secret, tlsServerCertsVolumeName, tlsServerCertsDir, tlsServerCertDirEnvVar, d)
-	}
-}
-
-func DeploymentRuntimeWithTLSClientSecret(secret string) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		mountTLSSecret(secret, tlsClientCertsVolumeName, tlsClientCertsDir, tlsClientCertDirEnvVar, d)
-	}
-}
-
-func DeploymentRuntimeWithAdditionalEnvironments(env []corev1.EnvVar) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Template.Spec.Containers[0].Env = append(d.Spec.Template.Spec.Containers[0].Env, env...)
-	}
-}
-
-func DeploymentRuntimeWithAdditionalPorts(ports []corev1.ContainerPort) DeploymentOverrides {
-	return func(d *appsv1.Deployment) {
-		d.Spec.Template.Spec.Containers[0].Ports = append(d.Spec.Template.Spec.Containers[0].Ports, ports...)
-	}
-}
-
-func (b *RuntimeManifestBuilder) Deployment(overrides ...DeploymentOverrides) *appsv1.Deployment {
-	d := b.defaultDeployment()
+func (b *RuntimeManifestBuilder) ServiceAccount(overrides ...ServiceAccountOverrides) *corev1.ServiceAccount {
+	sa := defaultServiceAccount(b.revision.GetName())
 
 	if b.controllerConfig != nil {
 		// Do something with the controller config
@@ -140,11 +72,39 @@ func (b *RuntimeManifestBuilder) Deployment(overrides ...DeploymentOverrides) *a
 	}
 
 	overrides = append(overrides,
-		DeploymentWithName(b.revision.GetName()),
+		// Currently it is not possible to override the namespace,
+		// ownerReferences or pullSecrets of the service account, and we could
+		// define them as defaults. However, we will leave them as overrides
+		// to indicate that we are opinionated about them currently and follow
+		// a consistent pattern.
+		ServiceAccountWithNamespace(b.namespace),
+		ServiceAccountWithOwnerReferences([]metav1.OwnerReference{meta.AsController(meta.TypedReferenceTo(b.revision, b.revision.GetObjectKind().GroupVersionKind()))}),
+		ServiceAccountWithAdditionalPullSecrets(append(b.revision.GetPackagePullSecrets(), b.serviceAccountPullSecrets...)),
+	)
+
+	for _, o := range overrides {
+		o(sa)
+	}
+
+	return sa
+}
+
+func (b *RuntimeManifestBuilder) Deployment(serviceAccount string, overrides ...DeploymentOverrides) *appsv1.Deployment {
+	d := defaultDeployment(b.revision.GetName())
+
+	if b.controllerConfig != nil {
+		// Do something with the controller config
+	}
+
+	if b.runtimeConfig != nil {
+		// Do something with the runtime config
+	}
+
+	overrides = append(overrides,
 		DeploymentWithNamespace(b.namespace),
 		DeploymentWithOwnerReferences([]metav1.OwnerReference{meta.AsController(meta.TypedReferenceTo(b.revision, b.revision.GetObjectKind().GroupVersionKind()))}),
 		DeploymentWithSelectors(b.podSelectors()),
-		DeploymentWithServiceAccount(b.revision.GetName()),
+		DeploymentWithServiceAccount(serviceAccount),
 		DeploymentWithImagePullSecrets(b.revision.GetPackagePullSecrets()),
 		DeploymentRuntimeWithImage(b.revision.GetSource()),
 	)
@@ -170,11 +130,6 @@ func (b *RuntimeManifestBuilder) Deployment(overrides ...DeploymentOverrides) *a
 	return d
 }
 
-func (b *RuntimeManifestBuilder) ServiceAccount() *corev1.ServiceAccount {
-	//TODO implement me
-	panic("implement me")
-}
-
 func (b *RuntimeManifestBuilder) Service() *corev1.Service {
 	//TODO implement me
 	panic("implement me")
@@ -194,44 +149,6 @@ func (b *RuntimeManifestBuilder) podSelectors() map[string]string {
 	return map[string]string{
 		"pkg.crossplane.io/revision":           b.revision.GetName(),
 		"pkg.crossplane.io/" + b.packageType(): b.packageName(),
-	}
-}
-
-func (b *RuntimeManifestBuilder) defaultDeployment() *appsv1.Deployment {
-	// TODO(turkenh): Implement configurable defaults.
-	// See https://github.com/crossplane/crossplane/issues/4699#issuecomment-1748403479
-	return &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Replicas: &replicas,
-			Template: corev1.PodTemplateSpec{
-				Spec: corev1.PodSpec{
-					SecurityContext: &corev1.PodSecurityContext{
-						RunAsNonRoot: &runAsNonRoot,
-						RunAsUser:    &runAsUser,
-						RunAsGroup:   &runAsGroup,
-					},
-					Containers: []corev1.Container{
-						{
-							Name:            runtimeContainerName,
-							ImagePullPolicy: corev1.PullIfNotPresent,
-							SecurityContext: &corev1.SecurityContext{
-								RunAsUser:                &runAsUser,
-								RunAsGroup:               &runAsGroup,
-								AllowPrivilegeEscalation: &allowPrivilegeEscalation,
-								Privileged:               &privileged,
-								RunAsNonRoot:             &runAsNonRoot,
-							},
-							Ports: []corev1.ContainerPort{
-								{
-									Name:          metricsPortName,
-									ContainerPort: metricsPortNumber,
-								},
-							},
-						},
-					},
-				},
-			},
-		},
 	}
 }
