+++ internal/controller/pkg/revision/runtime_defaults.go
@@ -1,0 +1,60 @@
+package revision
+
+import (
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func defaultServiceAccount(name string) *corev1.ServiceAccount {
+	return &corev1.ServiceAccount{
+		ObjectMeta: metav1.ObjectMeta{
+			// It is possible to override the name of the service account in the
+			// RuntimeConfig. So, we define it as a default here.
+			Name: name,
+		},
+	}
+}
+
+func defaultDeployment(name string) *appsv1.Deployment {
+	// TODO(turkenh): Implement configurable defaults.
+	// See https://github.com/crossplane/crossplane/issues/4699#issuecomment-1748403479
+	return &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			// It is possible to override the name of the deployment in the
+			// RuntimeConfig. So, we define it as a default here.
+			Name: name,
+		},
+		Spec: appsv1.DeploymentSpec{
+			Replicas: &replicas,
+			Template: corev1.PodTemplateSpec{
+				Spec: corev1.PodSpec{
+					SecurityContext: &corev1.PodSecurityContext{
+						RunAsNonRoot: &runAsNonRoot,
+						RunAsUser:    &runAsUser,
+						RunAsGroup:   &runAsGroup,
+					},
+					Containers: []corev1.Container{
+						{
+							Name:            runtimeContainerName,
+							ImagePullPolicy: corev1.PullIfNotPresent,
+							SecurityContext: &corev1.SecurityContext{
+								RunAsUser:                &runAsUser,
+								RunAsGroup:               &runAsGroup,
+								AllowPrivilegeEscalation: &allowPrivilegeEscalation,
+								Privileged:               &privileged,
+								RunAsNonRoot:             &runAsNonRoot,
+							},
+							Ports: []corev1.ContainerPort{
+								{
+									Name:          metricsPortName,
+									ContainerPort: metricsPortNumber,
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+}
